requirements.txt
flask
requests
bs4
ollama
pinecone-client
flask-cors
python-dotenv